refact: split apart app into routes + constants files; improve testing
